refactor: rename tests, docs: improved README.md.
