+++ CONTRIBUTING.md
@@ -1,0 +1,32 @@
+# Contributing rules
+
+Thank you for your help! Before you start, let's take a look at some agreements.
+
+## Pull request rules
+
+Make sure that your code:
+
+1.	Does not contain analyzer errors.
+2.	Follows a [official style](https://dart.dev/guides/language/effective-dart/style).
+3.  Follows the official [style of formatting](https://flutter.dev/docs/development/tools/formatting).
+4.	Contains no errors.
+5.	New functionality is covered by tests. New functionality passes old tests.
+6.	Create example that demonstrate new functionality if it is possible.
+
+## Accepting the changes
+
+After your pull request passes the review code, the project maintainers will merge the changes
+into the branch to which the pull request was sent.
+
+## Issues
+
+Feel free to report any issues and bugs.
+
+1.	To report about the problem, create an issue on GitHub.
+2.	In the issue add the description of the problem.
+3.	Do not forget to mention your development environment, Flutter version, libraries required for
+      illustration of the problem.
+4.	It is necessary to attach the code part that causes an issue or to make a small demo project
+      that shows the issue.
+5.	Attach stack trace so it helps us to deal with the issue.
+6.	If the issue is related to graphics, screen recording is required.
