+++ README.md
@@ -1,22 +1,21 @@
 # Keyboard Listener
 
-[![Build Status](https://shields.io/github/workflow/status/surfstudio/SurfGear/build?logo=github&logoColor=white)](https://github.com/surfstudio/SurfGear/tree/main/packages/keyboard_listener)
+[![Build Status](https://shields.io/github/workflow/status/surfstudio/flutter-bottom-inset-observer/Analysis?logo=github&logoColor=white)](https://github.com/surfstudio/flutter-bottom-inset-observer)
 
-This package made by [Surf](https://surf.ru).
+This package is part of the [SurfGear](https://github.com/surfstudio/SurfGear) toolkit made by [Surf](https://surf.ru).
 
 ## Description
 
-BottomInsetObserver - пакет для отслеживания изменений нижнего отступа (viewInsets.bottom) с помощью callBack функций.
+BottomInsetObserver - a package for large indent changes (viewInsets.bottom) using the callBack function.
 
 ## Installation
 
-Add `keyboard_listener` to your `pubspec.yaml` file:
+Add `bottom_inset_observer` to your `pubspec.yaml` file:
+
 
 ```yaml
 dependencies:
-  keyboard_listener:
-    git:
-      url: git://github.com/surfstudio/flutter-keyboard-listener.git
+  bottom_inset_observer: $currentVersion$
 ```
 
 ## Example
@@ -103,11 +102,11 @@ All notable changes to this project will be documented in [this file](./CHANGELO
 
 ## Issues
 
-For issues, file directly in the Issues section.
+To report your issues, submit them directly in the [Issues](https://github.com/surfstudio/flutter-bottom-inset-observer/issues) section.
 
 ## Contribute
 
-If you would like to contribute to the package (e.g. by improving the documentation, solving a bug or adding a cool new feature), please review our [contribution guide](../../CONTRIBUTING.md) first and send us your pull request.
+If you would like to contribute to the package (e.g. by improving the documentation, fixing a bug or adding a cool new feature), please review our [contribution guide](/CONTRIBUTING.md) first and send us your pull request.
 
 Your PRs are always welcome.
